+++ Sources/Remote Logging/Crash Logging/CrashLogging.swift
@@ -146,6 +146,22 @@ public class CrashLogging {
 // MARK: - Manual Error Logging
 public extension CrashLogging {
 
+    /// Writes a JavaScript exception to the Crash Logging system, including its stack trace.
+    /// Note that this function is provided mainly for hybrid sources like React Native.
+    ///
+    /// - Parameters:
+    ///   - exception: The exception object
+    ///   - callback: Callback triggered upon completion
+    func logJavaScriptException(_ jsException: any JSException, callback: @escaping () -> Void) {
+        
+        SentrySDK.capture(event: SentryEventJSException.init(jsException: jsException))
+        
+        DispatchQueue.global().async {
+            SentrySDK.flush(timeout: self.flushTimeout)
+            callback()
+        }
+    }
+    
     /// Writes the error to the Crash Logging system, and includes a stack trace. This API supports for rich events.
     /// By setting a Tag/Value pair, you'll be able to filter these events, directly, with the `has:` operator (Sentry Web Interface).
     ///